refactor: split redis & mongo to single file
